+++ CHANGELOG.md
@@ -2,23 +2,45 @@
 
 All notable changes to this project will be documented in this file.
 
+## [???] - 2020-05-04
+
+### Added
+
+- standardized RF documentation file generated via libdoc
+- added .gitignore file
+- added LICENCE file
+
+### Changed
+
+- unittests moved to new location, separated from WatchUI library file
+- unittests .robot file changed to reflect the move
+- WatchUI.py docstrings modified to generate documentation via libdoc
+- minor refactoring of WatchUI.py to remove redundant variable assignments
+- modified unittests github pipeline to reflect unittests move
+
 ## [0.17.8] - 2020-04-10
 
 ### Added
+
 - Created new unit test for starting CI
 - Added screen in robot framework log if the test is pass
 - Created Actions on github.com
+
 ### Changed
+
 - Change in code (more OOP)
 
 ## [0.17.2] - 2020-03-25
 
 ### Added
-- SSIM (Threshold) and outputs folder can be set in *** settings *** (Library output_foder= ssim=)
+
+- SSIM (Threshold) and outputs folder can be set in **_ settings _** (Library output_foder= ssim=)
 - With settings we must change code, so now you can find class in ours code :-)
 - Keyword Create Area has screen_name now
 - Every keyword for compare has settings for lowest limit of difference. ssim=
+
 ### Changed
+
 - Changed outputs folder from name by keyword name, just on ../Outputs
 - Repair bug in keyword Compare screen get info. Now if you set bad path, give you error
 - Change ssim from int to float
@@ -27,12 +49,13 @@ All notable changes to this project will be documented in this file.
 ## [0.16.4] - 2020-03-19
 
 ### Changed
+
 - Change keyword `Compare making rescreens` on `Create sreens`
 - Change keyword `Compare making are` on `Create area`
-- Change name folder in keyword Create screens from `name_screen` on `screen_name` 
+- Change name folder in keyword Create screens from `name_screen` on `screen_name`
 - Fixed keyword in test.robot
 - Some change in code
-- Add new creator  Radek Bednařík in code, thanks for help :-)
+- Add new creator Radek Bednařík in code, thanks for help :-)
 - Fixed bugs in area keywords
 - After install with pypi u can using with full apth your watchUI in robot framework
 
@@ -42,8 +65,8 @@ All notable changes to this project will be documented in this file.
 
 - Add settings for Compare making rescreens. Now you can set up name for folder and picture.
 
-
 ### Changed
+
 - Fixed some log to robot
 - Fixed readme
 - Fixed error in test.robot (Library Watchmen -> Library WatchUI)